Dump of beginning of rust barcode scanner listening thing.
Have test.c which uses libevdev to print out events that it reads from a
given input.

Have main.rs with the extern definitions that will be needed to
interface with libevdev from rust.

TODO: Replicate the exact same behavior in rust
TODO: Take the target device as an arg